Make tests pass in case of unknown compiler in jenkins
Make the tests pass if the compiler is unknown and the travis flag is
passed by short circuiting the build and run processes for all languages.

This workaround would not have been necessary if the testing infrastructure
provided provisions for supporting different configurations for different
branches on the repository. Remove this workaround once test infrastructure
supports configurations for different branches.

TODO (https://github./grpc/grpc/issues/7934)