Modify run_tests to pass for unknown compilers
Default compiler to Python2.7 in case of unknown compiler
if the travis flag is present. This is needed as a workaround for issue #7934
This wouldn't have been necessary if the testing infrastructure provided
a way to provision different configuration for different repository
branches.

TODO (https://github.com/grpc/grpc/issues/7934)